feature: separate exam from question
