+++ exam/tests.py
@@ -9,11 +9,6 @@ Unless required by applicable Law or agreed to in writing, software distributed 
 an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific Language governing permissions and limitations under the License.
 """
-from django.contrib import admin
+# from django.test import TestCase
 
-from .models import Question, QuestionTag, QuestionTagContact
-
-# Register your models here.
-admin.site.register(Question)
-admin.site.register(QuestionTag)
-admin.site.register(QuestionTagContact)
+# Create your tests here.
